Move ListOp into its own module
